--- js/components/big-picture.js
@@ -1,97 +1,0 @@
-const body = document.body;
-const bigPicture = body.querySelector('.big-picture');
-const commentCount = bigPicture.querySelector('.social__comment-count');
-const commentList = bigPicture.querySelector('.social__comments');
-const commentsLoader = bigPicture.querySelector('.comments-loader');
-
-const cancelButton = bigPicture.querySelector('.big-picture__cancel');
-
-const COMMENTS_PER_PORTION = 5;
-let commentsShown = 0;
-let comments = [];
-
-const createComment = ({ avatar, name, message }) => {
-  const comment = document.createElement('li');
-  comment.innerHTML =
-    '<img class="social__picture" src="" alt="" width="35" height="35"><p class="social__text"></p>';
-  comment.classList.add('social__comment');
-
-  const image = comment.querySelector('.social__picture');
-  const text = comment.querySelector('.social__text');
-
-  image.src = avatar;
-  image.alt = name;
-  text.textContent = message;
-
-  return comment;
-};
-
-const renderComments = () => {
-  commentsShown += COMMENTS_PER_PORTION;
-
-  if (commentsShown >= comments.length) {
-    commentsLoader.classList.add('hidden');
-    commentsShown = comments.length;
-  } else {
-    commentsLoader.classList.remove('hidden');
-  }
-
-  const fragment = document.createDocumentFragment();
-  for (let i = 0; i < commentsShown; i++) {
-    const commentElement = createComment(comments[i]);
-    fragment.append(commentElement);
-  }
-
-  commentList.innerHTML = '';
-  commentList.append(fragment);
-  commentCount.innerHTML = `${commentsShown} из <span class="comments-count">${comments.length}</span> комментариев`;
-};
-
-const hideBigPicture = () => {
-  bigPicture.classList.add('hidden');
-  body.classList.remove('modal-open');
-  document.removeEventListener('keydown', onEscKeyDown);
-  commentsShown = 0;
-};
-
-function onEscKeyDown(evt) {
-  if (evt.key === 'Escape') {
-    evt.preventDefault();
-    hideBigPicture();
-  }
-}
-
-const onCancelButtonClick = () => {
-  hideBigPicture();
-};
-
-const onCommentsLoaderClick = () => renderComments();
-
-const renderPictureDetails = ({ url, likes, description }) => {
-  const img = bigPicture.querySelector('.big-picture__img img');
-  const likesCount = bigPicture.querySelector('.likes-count');
-  const social = bigPicture.querySelector('.social__caption');
-
-  img.src = url;
-  img.alt = description;
-  likesCount.textContent = likes;
-  social.textContent = description;
-};
-
-const showBigPicture = (data) => {
-  bigPicture.classList.remove('hidden');
-  body.classList.add('modal-open');
-  commentsLoader.classList.add('hidden');
-  document.addEventListener('keydown', onEscKeyDown);
-
-  renderPictureDetails(data);
-  comments = data.comments;
-  if (comments.length > 0) {
-    renderComments();
-  }
-};
-
-cancelButton.addEventListener('click', onCancelButtonClick);
-commentsLoader.addEventListener('click', onCommentsLoaderClick);
-
-export { showBigPicture };
