+++ js/components/render-pictures.js
@@ -1,4 +1,4 @@
-import { showBigPicture } from './big-picture';
+import { showBigPicture } from './show-big-picture';
 
 const pictureTemplate = document.querySelector('#picture')
   .content.querySelector('.picture');
